+++ .pre-commit-config.yaml
@@ -14,3 +14,8 @@ repos:
     rev: 22.12.0
     hooks:
     -   id: black
+-   repo: https://github.com/pycqa/isort
+    rev: 5.12.0
+    hooks:
+      - id: isort
+        name: isort (python)
