+++ models/common.py
@@ -23,25 +23,18 @@
 # AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.from models.models import BaseModel, DumpedModel
+# SOFTWARE.
+
+from typing import Any, TypeAlias, TypeVar
+
+from sqlalchemy.ext.asyncio import AsyncAttrs
+from sqlalchemy.orm import DeclarativeBase
 
 
-from sqlalchemy import Integer, String
-from sqlalchemy.orm import Mapped, mapped_column
-
-from models.base_models import BaseModel, DumpedModel
+class BaseModel(AsyncAttrs, DeclarativeBase):
+    pass
 
 
-class Permission(BaseModel):
-    __tablename__ = "permissions"
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    permission: Mapped[str] = mapped_column(String(), nullable=False)
-    description: Mapped[str | None] = mapped_column(String())
-
-    # Format method
-    def format(self, **kwargs) -> DumpedModel:
-        return {
-            "id": self.id,
-            "permission": self.permission,
-            "description": self.description,
-        }
+HugModelType = TypeVar("HugModelType", bound=BaseModel, covariant=True)
+DumpedModel: TypeAlias = dict[str, Any]
+BLOCKED_USER_ROLE_ID = 5
