+++ setup.cfg
@@ -8,3 +8,18 @@ addopts =
 	--cov-config=.coveragerc
 	--cov=.
 asyncio_mode = auto
+
+[isort]
+# https://github.com/timothycrosley/isort
+# https://github.com/timothycrosley/isort/wiki/isort-Settings
+profile = black
+# Any local folders we import from have to be entered here
+known_local_folder =
+    models,
+    utils,
+force_sort_within_sections = true
+include_trailing_comma=True
+use_parentheses=True
+sections = FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
+combine_as_imports = true
+indent = "    "
